utils ts migration
